// clients/accountingService/models/TransactionCategoryUpdateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TransactionCategoryUpdateDto = {
    name?: string | null;
    description?: string | null;
    enrolmentId?: string | null;
    tenantId?: string | null;
};