// clients/accountingService/models/TransactionCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TransactionCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    description?: string | null;
    price?: number;
    quantity?: number;
    externalDescription?: string | null;
    basisQuantity?: number;
    basisAmount?: number;
    percent?: number;
    unitGroupId?: string | null;
    unitId?: string | null;
    transactionCategoryId?: string | null;
    currencyId?: string | null;
    tenantId?: string | null;
    enrolmentId?: string | null;
};